make basic auth config immutable
