+++ pkg/config/config.go
@@ -9,13 +9,6 @@ import (
 	"github.com/kelseyhightower/envconfig"
 )
 
-// BasicAuthConfig represents details such as username and password that will
-// be used to secure the broker using basic auth
-type BasicAuthConfig struct {
-	Username string `envconfig:"BASIC_AUTH_USERNAME" required:"true"`
-	Password string `envconfig:"BASIC_AUTH_PASSWORD" required:"true"`
-}
-
 // ModulesConfig represents details re: which modules should be included or
 // excluded when the broker is started
 type ModulesConfig struct {
@@ -34,13 +27,6 @@ type AzureConfig struct {
 	ClientSecret         string `envconfig:"AZURE_CLIENT_SECRET" required:"true"`
 	DefaultLocation      string `envconfig:"AZURE_DEFAULT_LOCATION"`
 	DefaultResourceGroup string `envconfig:"AZURE_DEFAULT_RESOURCE_GROUP"`
-}
-
-// GetBasicAuthConfig returns basic auth configuration
-func GetBasicAuthConfig() (BasicAuthConfig, error) {
-	bac := BasicAuthConfig{}
-	err := envconfig.Process("", &bac)
-	return bac, err
 }
 
 // GetModulesConfig returns modules configuration
